% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{accumulate_ADMnew_Patch}
\alias{accumulate_ADMnew_Patch}
\title{Accumulate ADMnew}
\usage{
accumulate_ADMnew_Patch(count)
}
\arguments{
\item{count}{vector of new ADM males}
}
\description{
Accumulate new ADM males
}
