% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{AnalyzeQuantiles}
\alias{AnalyzeQuantiles}
\title{Summary Statistics for Stochastic MGDrivE}
\usage{
AnalyzeQuantiles(readDirectory, writeDirectory, mean = TRUE,
  quantiles = NULL)
}
\arguments{
\item{readDirectory}{Directory to find repetition folders in}

\item{writeDirectory}{Directory to write output}

\item{mean}{Boolean, calculate mean or not. Default is TRUE}

\item{quantiles}{Vector of quantiles to calculate. Default is NULL}
}
\value{
Writes output to files in writeDirectory
}
\description{
This function reads in all repetitions for each patch and calculates either
the mean, quantiles, or both. User chooses the quantiles, up to 4 decimal places,
and enters them as a vector. (order does not matter)  \cr
}
\details{
Given the readDirectory, this function assumes the follow file structure: \cr
\itemize{
\item readDirectory
\itemize{
\item repetition 1
\itemize{
\item patch 1
\item patch 2
\item patch 3
}
\item repetition 2
\itemize{
\item patch 1
\item patch 2
\item patch 3
}
\item repetition 3
\item repetition 4
\item ...
}
}
}
