% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_ExpKernel}
\alias{calc_ExpKernel}
\title{Calculate Exponential Stochastic Matrix}
\usage{
calc_ExpKernel(distMat, r)
}
\arguments{
\item{distMat}{distance matrix from \code{\link[MGDrivE]{calc_haversine}}}

\item{r}{rate parameter of \code{\link[stats]{Exponential}} distribution}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calc_haversine}}, calculate a stochastic matrix where one step movement probabilities follow an exponential density.
}
