% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_GammaKernel}
\alias{calc_GammaKernel}
\title{Calculate Gamma Stochastic Matrix}
\usage{
calc_GammaKernel(distMat, shape, rate)
}
\arguments{
\item{distMat}{distance matrix from \code{\link[MGDrivE]{calc_haversine}}}

\item{shape}{shape parameter of \code{\link[stats]{GammaDist}} distribution}

\item{rate}{rate parameter of \code{\link[stats]{GammaDist}} distribution}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calc_haversine}}, calculate a stochastic matrix where one step movement probabilities follow a gamma density.
}
