% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_haversine}
\alias{calc_haversine}
\title{Calculate Haversine Distance}
\usage{
calc_haversine(longlats)
}
\arguments{
\item{longlats}{numeric matrix where first column is vector of longitudes and second column is vector of latitudes}
}
\description{
Calculate the great-circle distance (Haversine distance) between sets of longitude - latitude points, see \url{https://en.wikipedia.org/wiki/Haversine_formula}
}
