% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_HurdleExpKernel}
\alias{calc_HurdleExpKernel}
\title{Calculate Hurdle Exponential Stochastic Matrix}
\usage{
calc_HurdleExpKernel(distMat, r, pi)
}
\arguments{
\item{distMat}{distance matrix from \code{\link[MGDrivE]{calc_haversine}}}

\item{r}{rate parameter of \code{\link[stats]{Exponential}} distribution}

\item{pi}{point mass at zero}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calc_haversine}}, calculate a stochastic matrix where one step movement probabilities follow an zero-truncated exponential density with a point mass at zero.
}
