% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_LognormalKernel}
\alias{calc_LognormalKernel}
\title{Calculate Lognormal Stochastic Matrix}
\usage{
calc_LognormalKernel(distMat, meanlog, sdlog)
}
\arguments{
\item{distMat}{distance matrix from \code{\link[MGDrivE]{calc_haversine}}}

\item{meanlog}{log mean of \code{\link[stats]{Lognormal}} distribution}

\item{sdlog}{log standard deviation of \code{\link[stats]{Lognormal}} distribution}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calc_haversine}}, calculate a stochastic matrix where one step movement probabilities follow a lognormal density.
}
