% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcAquaticStagesSurvivalProbability}
\alias{calcAquaticStagesSurvivalProbability}
\title{Calculate Survival Probability of entire Aquatic Stage Life-cycle}
\usage{
calcAquaticStagesSurvivalProbability(eggSurvivalProbability,
  larvaSurvivalProbability, pupaSurvivalProbability)
}
\arguments{
\item{eggSurvivalProbability}{see \code{\link{calcAquaticStageSurvivalProbability}}}

\item{larvaSurvivalProbability}{see \code{\link{calcAquaticStageSurvivalProbability}}}

\item{pupaSurvivalProbability}{see \code{\link{calcAquaticStageSurvivalProbability}}}
}
\description{
Calculate vector of survival probabilities for each stage of aquatic life-cycle.
}
