% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcAquaticStageSurvivalProbability}
\alias{calcAquaticStageSurvivalProbability}
\title{Calculate Aquatic Stage Survival Probability}
\usage{
calcAquaticStageSurvivalProbability(mortalityRate, stageDuration)
}
\arguments{
\item{mortalityRate}{Daily mortality probability, \eqn{\mu_{st}}}

\item{stageDuration}{Duration of aquatic stage, \eqn{T^{st}}}
}
\description{
Calculate \eqn{\theta_{st}}, density-independent survival probability, given
by: \deqn{\theta_{st}=(1-\mu_{st})^{T_{st}}}
}
