% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcDensityDependentDeathRate}
\alias{calcDensityDependentDeathRate}
\title{Calculate Density-dependent Adult Mortality}
\usage{
calcDensityDependentDeathRate(
  fertility,
  litters,
  thetaAq,
  thetaAd,
  tAq,
  carryingCap,
  populationGrowthRate
)
}
\arguments{
\item{fertility}{Number of pups per litter for wild-type females, \eqn{\beta}}

\item{litters}{number of litters per female mouse per year}

\item{thetaAq}{Vector of density-independent survival probabilities of adolescent stages, \eqn{\theta_{e}, \theta_{l}}}

\item{thetaAd}{density-independent survival probability of adult stage}

\item{tAq}{Length of Adolescent Life Stages, \eqn{T_{a}}}

\item{carryingCap}{Adult population carrying capacity, \eqn{k}}

\item{populationGrowthRate}{Population growth in absence of density-dependent mortality \eqn{R_{m}}}
}
\description{

}
