% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{calcLarvalDist}
\alias{calcLarvalDist}
\title{Calculate Distribution of Larval Population}
\usage{
calcLarvalDist(mu, t)
}
\arguments{
\item{mu}{Double, death rate}

\item{t}{Integer, stage time}
}
\description{
This hidden function calculates the distribution of larvae through time by
treating the larval-stage as a discrete-time Markov chain, and solving for the
stationary distribution. As the only aquatic population known for initializing
MGDrivE is the equilibrium larval population, this acts as an anchor from which
to calculate the egg and pupae distributions from (see \code{\link{set_initialPopulation_Patch}}).
}
