% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcLarvalPopEquilibrium}
\alias{calcLarvalPopEquilibrium}
\title{Calculate Equilibrium Larval Population}
\usage{
calcLarvalPopEquilibrium(carryingCap, muAq)
}
\arguments{
\item{carryingCap}{See \code{\link{calcDensityDependentDeathRate}}}

\item{muAq}{daily mortalities of each adolescent life stage}
}
\description{
Equilibrium larval population size to sustain adult population.
}
