% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcPopulationGrowthRate}
\alias{calcPopulationGrowthRate}
\title{Calculate Generational Population Growth Rate}
\usage{
calcPopulationGrowthRate(dailyPopGrowthRate, averageGenerationTime)
}
\arguments{
\item{dailyPopGrowthRate}{Daily population growth rate, \eqn{r_{m}}}

\item{averageGenerationTime}{See \code{\link{calcAverageGenerationTime}}}
}
\description{
Calculate \eqn{R_{m}}, population growth in absence of density-dependent mortality,
given by: \deqn{(r_{m})^{g}}
}
