% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{close_allConnections_Network}
\alias{close_allConnections_Network}
\title{Close all Output Connections}
\usage{
close_allConnections_Network()
}
\description{
Close \code{private$conADM} and \code{private$conAF1}
}
