% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{createNamedPopMatrix}
\alias{createNamedPopMatrix}
\title{Create a Named Matrix}
\usage{
createNamedPopMatrix(genotypesID)
}
\arguments{
\item{genotypesID}{character vector of possible genotypes}
}
\description{
Create a named matrix of 0s
}
