% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{createNamedPopVector}
\alias{createNamedPopVector}
\title{Create a Named Vector}
\usage{
createNamedPopVector(genotypesID)
}
\arguments{
\item{genotypesID}{character vector of possible genotypes}
}
\description{
Create a named vector of 0s
}
