% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube-CRISPR1RA.R
\name{Cube_Homing1RA}
\alias{Cube_Homing1RA}
\title{Inheritance Cube: Homing Drive with 1 Resistance Allele}
\usage{
Cube_Homing1RA(e = 1, p = 0, eta = NULL, phi = NULL, omega = NULL,
  xiF = NULL, xiM = NULL, s = NULL)
}
\arguments{
\item{e}{Homing rate}

\item{p}{Resistance allele generation rate}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list(inheritance cube, viability mask, genotypes ID, genotypes number,
wild-type allele, mating fitness, sex ratio, adult mortality modifier, female pupatory success,
male pupatory success, fertility modifier, release genotype)
}
\description{
This function creates an inheritance cube to model a homing gene drive (such as a CRISPR-Cas9 system)
that creates 1 type of resistance allele. It assumes no sex-specific inheritance patterns and the
construct is on an autosome.
}
