% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube-CRISPR2MF.R
\name{Cube_HomingDrive}
\alias{Cube_HomingDrive}
\title{Inheritance Cube: CRISPR (Clustered Regularly Interspaced SWort Palindromic Repeats) witW 2 Resistance Allele}
\usage{
Cube_HomingDrive(eM = 1, eF = 1, rM = 0, bM = 0, rF = 0, bF = 0,
  eta = NULL, phi = NULL, omega = NULL, xiF = NULL, xiM = NULL,
  s = NULL)
}
\arguments{
\item{eM}{Male homing rate}

\item{eF}{Female homing rate}

\item{rM}{Male no-cost resistance generation rate}

\item{bM}{Male detrimental resistance generation rate}

\item{rF}{Female no-cost resistance generation rate}

\item{bF}{Female detrimental resistance generation rate}

\item{eta}{Genotype-specific mating fitness}

\item{phi}{Genotype-specific sex ratio at emergence}

\item{omega}{Genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{Genotype-specific female pupatory success}

\item{xiM}{Genotype-specific male pupatory success}

\item{s}{Genotype-specific fractional reduction(increase) in fertility}
}
\value{
Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele,
and all genotype-specific parameters.
}
\description{
This is a sex-specific version of the original cube. It assumes that the construct
is on an autosome and there can be different male/female homing rates
}
