% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube-JOHNConfineableHomingWithRandB.R
\name{cubeConfinableHomingJOHN}
\alias{cubeConfinableHomingJOHN}
\title{Inheritance Cube: Confinable Homing Drive, John}
\usage{
cubeConfinableHomingJOHN(
  eM = 1,
  eF = 1,
  prRF = 0,
  prRM = 0,
  r = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{eM}{Male homing rate}

\item{eF}{Female homing rate}

\item{prRF}{Female no-cost resistant allele generation rate}

\item{prRM}{Male no-cost resistant allele generation rate}

\item{r}{Crossover probability}

\item{eta}{genotype-specific mating fitness}

\item{phi}{genotype-specific sex ratio at emergence}

\item{omega}{genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{genotype-specific female pupatory success}

\item{xiM}{genotype-specific male pupatory success}

\item{s}{genotype-specific fractional reduction(increase) in fertility}
}
\description{
write me
}
