% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube-Auxiliary.R
\name{cubeModifiers}
\alias{cubeModifiers}
\title{Generate and Modify Default Genotype-specific Parameters}
\usage{
cubeModifiers(
  gtype,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)
}
\arguments{
\item{gtype}{character vector of genotypes}

\item{eta}{genotype-specific mating fitness, handles assortative mating as well}

\item{phi}{genotype-specific sex ratio at emergence}

\item{omega}{genotype-specific multiplicative modifier of adult mortality}

\item{xiF}{genotype-specific female pupatory success}

\item{xiM}{genotype-specific male pupatory success}

\item{s}{genotype-specific fractional reduction(increase) in fertility}
}
\description{
This is an internal function for cubes.
}
