% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{eraseDirectory}
\alias{eraseDirectory}
\title{Erase all files in a directory}
\usage{
eraseDirectory(directory)
}
\arguments{
\item{directory}{directory whose contents will be deleted}
}
\description{
Given a directory path, check it exists and if so delete all its contents.
}
