% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Releases.R
\name{generateReleaseVector}
\alias{generateReleaseVector}
\title{Make List of Modified Mosquito Releases}
\usage{
generateReleaseVector(driveCube = driveCube,
  releasesParameters = releasesParameters, sex = "M")
}
\arguments{
\item{driveCube}{gene-drive cube}

\item{releasesParameters}{A list containing the releasesStart, releasesNumber
releasesInterval, and releaseProportion named values.}

\item{sex}{character in 'M','F'}
}
\description{
Sets up a release schedule for a single patch, calls \code{\link{Release_basicRepeatedReleases}} internally.
}
