% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_alpha_Network}
\alias{get_alpha_Network}
\title{Get alpha}
\usage{
get_alpha_Network(ix)
}
\arguments{
\item{ix}{Index of patch}
}
\description{
Return density dependent mortality, see \code{\link{calcDensityDependentDeathRate}}
}
