% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_batchLocRow_Network}
\alias{get_batchLocRow_Network}
\title{Get Batch Location Distribution}
\usage{
get_batchLocRow_Network(ix)
}
\arguments{
\item{ix}{Index of patch}
}
\description{
Return the distribution of location probabilities
}
