% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_batchProbs_Network}
\alias{get_batchProbs_Network}
\title{Get Batch Migration Probability}
\usage{
get_batchProbs_Network(ix)
}
\arguments{
\item{ix}{index of patch}
}
\description{
Return the probability of undergoing batch migration each day
}
