% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_batchSex_Network}
\alias{get_batchSex_Network}
\title{Get Batch Migration Sex}
\usage{
get_batchSex_Network(ix, sex)
}
\arguments{
\item{ix}{index of patch}

\item{sex}{number, M=1 and F=2}
}
\description{
Return the batch migration size for each sex
}
