% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_driveCube_genotype_Network}
\alias{get_driveCube_genotype_Network}
\title{Get Element(s) of Drive Cube by Genotype}
\usage{
get_driveCube_genotype_Network(fG = NULL, mG = NULL, oG = NULL)
}
\arguments{
\item{fG}{female genotype}

\item{mG}{male genotype}

\item{oG}{offspring genotype}
}
\description{
Return elements or slices of drive cube. If all \code{NULL} return entire cube.
}
