% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_driveCube_index_Network}
\alias{get_driveCube_index_Network}
\title{Get Element(s) of Drive Cube by Index}
\usage{
get_driveCube_index_Network(fG = NULL, mG = NULL, oG = NULL)
}
\arguments{
\item{fG}{female genotype index}

\item{mG}{male genotype index}

\item{oG}{offspring genotype index}
}
\description{
Return elements or slices of drive cube. If all \code{NULL} return entire cube.
}
