% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_drivecubeindex_Network}
\alias{get_drivecubeindex_Network}
\title{Get Element(s) of Drive Cube by Index}
\usage{
get_drivecubeindex_Network(fG = NULL, mG = NULL, oG = NULL)
}
\arguments{
\item{fG}{Female genotype index}

\item{mG}{Male genotype index}

\item{oG}{Offspring genotype index}
}
\description{
Return elements or slices of drive cube. If all \code{NULL} return entire cube.
}
