% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_eta_Network}
\alias{get_eta_Network}
\title{Get eta}
\usage{
get_eta_Network(fIdx)
}
\arguments{
\item{fIdx}{Index of female genotype to pull}
}
\description{
Get eta
}
\keyword{Return}
\keyword{fitness}
\keyword{genotype-specific}
\keyword{internal}
\keyword{mating}
