% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{get_femaleMigration_Patch}
\alias{get_femaleMigration_Patch}
\title{Get maleMigration}
\usage{
get_femaleMigration_Patch()
}
\description{
Return outbound males (nGenotypes X nGenotypes X nPatch array)
}
