% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_genPopGrowth_Network}
\alias{get_genPopGrowth_Network}
\title{Get genPopGrowth}
\usage{
get_genPopGrowth_Network()
}
\description{
Return population growth rate, see \code{\link{calcPopulationGrowthRate}}
}
