% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_Leq_Network}
\alias{get_Leq_Network}
\title{Get Leq}
\usage{
get_Leq_Network(ix)
}
\arguments{
\item{ix}{index of patch}
}
\description{
Return equilibrium larval population, see \code{\link{calcLarvalPopEquilibrium}}
}
