% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_migrationMaleRow_Network}
\alias{get_migrationMaleRow_Network}
\title{Get Row of Male Migration Matrix}
\usage{
get_migrationMaleRow_Network(ix)
}
\arguments{
\item{ix}{index of row}
}
\description{
Return a matrix object (does not drop dimensions)
}
