% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{get_NetworkPointer_Patch}
\alias{get_NetworkPointer_Patch}
\title{Get Network Pointer}
\usage{
get_NetworkPointer_Patch()
}
\description{
Return a reference to the enclosing \code{\link{Network}} object
}
