% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_patch_Network}
\alias{get_patch_Network}
\title{Get Patch}
\usage{
get_patch_Network(ix)
}
\arguments{
\item{ix}{integer id of patch to return}
}
\description{
Return a \code{\link{Patch}} object
}
