% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_theta_Network}
\alias{get_theta_Network}
\title{Get theta}
\usage{
get_theta_Network()
}
\arguments{
\item{ix}{Index of patch}
}
\description{
Return shape parameter for carrying capacity function
}
