% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_thetaAq_Network}
\alias{get_thetaAq_Network}
\title{Get thetaAq}
\usage{
get_thetaAq_Network(stage)
}
\arguments{
\item{stage}{character in 'E', 'L', 'P'}
}
\description{
Return aquatic stage survival probability, see \code{\link{calcAquaticStagesSurvivalProbability}} and \code{\link{calcAquaticStageSurvivalProbability}}
}
