% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_timeAq_Network}
\alias{get_timeAq_Network}
\title{Get timeAq}
\usage{
get_timeAq_Network(stage = NULL)
}
\arguments{
\item{stage}{Character in 'E', 'L', 'P'; if \code{NULL} return total duration}
}
\description{
Return duration of aquatic stages.
}
