% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_timeJu_Network}
\alias{get_timeJu_Network}
\title{Get timeJu}
\usage{
get_timeJu_Network(stage = NULL)
}
\arguments{
\item{stage}{Character in 'G', 'N', 'A'; if \code{NULL} return total duration}
}
\description{
Return duration of juvenile stages.
}
