% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Graphics.R
\name{ggCol_utility}
\alias{ggCol_utility}
\title{Utility to Imitate ggplot2 Colors}
\usage{
ggCol_utility(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transparency}
}
\description{
Sample at equally spaced intervals along the color wheel
}
