% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Graphics.R
\name{ggColUtility}
\alias{ggColUtility}
\title{Utility to Imitate ggplot2 Colors}
\usage{
ggColUtility(n, alpha = 0.75)
}
\arguments{
\item{n}{Number of colors}

\item{alpha}{Transparency}
}
\description{
Sample at equally spaced intervals along the color wheel
}
