% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{initPopMatrixArray}
\alias{initPopMatrixArray}
\title{Create a population array of matrices}
\usage{
initPopMatrixArray(genotypesID, memoryWindow)
}
\arguments{
\item{genotypesID}{character vector of possible genotypes}

\item{memoryWindow}{integer size of list structure}
}
\description{
Creates an array for the population history to be stored. The length of the
array is equal to the window required for the model to run
(in our specific case it is equal to the sum of aquatic stages lengths).
}
