% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{initStagesDurations}
\alias{initStagesDurations}
\title{Initialize Aquatic Stages Durations}
\usage{
initStagesDurations(egg = 1, larva = 14, pupa = 1)
}
\arguments{
\item{egg}{length of egg stage (days)}

\item{larva}{length of larval stage (days)}

\item{pupa}{length of pupal stage (days)}
}
\description{
Initialises the vector that holds the duration of each aquatic stage
}
