% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kernels}
\alias{kernels}
\title{Kernels Parameters}
\format{
named list with 5 elements:
\describe{
\item{lnorm_mean}{log mean of log-normal density}
\item{lnorm_sd}{log standard deviation of log-normal density}
\item{gamma_shape}{shape parameter of gamma density}
\item{gamma_sd}{rate parameter of gamma density}
\item{exp_rate}{rate parameter of exponential density}
}
}
\usage{
data(kernels)
}
\description{
A named list containing maximum likelihood fitted parameter values from
mosquito dispersal estimates.
}
\keyword{datasets}
