% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDextract.R
\name{MGDcpt}
\alias{MGDcpt}
\title{Function by: Ethan A. Brown (March 4 2020)}
\usage{
MGDcpt(
  out_path. = out_path,
  tpoints = tMax,
  tMax. = tMax,
  cube. = cube,
  nrep. = nRep,
  sitesNumber. = sitesNumber
)
}
\arguments{
\item{out_path.}{complete file path to the output folder (where the simulation directories are located)}

\item{tpoints}{numeric vector of the specific time points you are interested in}

\item{tMax.}{numeric vector of the specific time points you are interested in}

\item{cube.}{inheritance cube object (same one you used to run the model)}

\item{nrep.}{number of simulations}

\item{sitesNumber.}{number of patches}
}
\description{
Consolidates output of MGDrivE into a conditional probability table for use in a Bayesian network
}
