% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Setup.R
\name{MGDrivE.Setup}
\alias{MGDrivE.Setup}
\title{Setup MGDrivE}
\usage{
MGDrivE.Setup(stochasticityON = FALSE)
}
\arguments{
\item{stochasticityON}{enable/disable stochastic simulation.}
}
\description{
Initialize methods in \code{\link{Patch}} to run deterministic or stochastic simulations.
}
