% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatAll2}
\alias{moveMatAll2}
\title{Movement Matrix: All 2}
\format{
A matrix with 3 rows and 3 columns:
\describe{
Patches 1 and 3 are sources for patch 2, which is a sink.
}
}
\usage{
data(moveMatAll2)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
