% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatCascade3}
\alias{moveMatCascade3}
\title{Movement Matrix: Cascade 3}
\format{
A matrix with 3 rows and 3 columns:
\describe{
Mosquitoes in patch 1 have equal probability to stay or move to 2; mosquitoes
in patch 2 have equal probability to stay or move to 3; mosquitoes in patch 3 stay there.
}
}
\usage{
data(moveMatCascade3)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
