% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatDiag}
\alias{moveMatDiag}
\title{Movement Matrix: Diagonal}
\format{
A matrix with 3 rows and 3 columns:
\describe{
3 independent patches.
}
}
\usage{
data(moveMatDiag)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
