% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatDiagOneCity}
\alias{moveMatDiagOneCity}
\title{Movement Matrix: Diagonal One City}
\format{
A matrix with 1 rows and 1 columns:
\describe{
A 1 by 1 matrix with entry 1.
}
}
\usage{
data(moveMatDiagOneCity)
}
\description{
A movement matrix for simulation with 1 patch.
}
\keyword{datasets}
