% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatDie}
\alias{moveMatDie}
\title{Movement Matrix: Die}
\format{
A matrix with 3 rows and 3 columns:
\describe{
All entries of matrix are 0 for testing that all mosquitoes will be killed.
}
}
\usage{
data(moveMatDie)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
