% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatIndependent3}
\alias{moveMatIndependent3}
\title{Movement Matrix: Independent 3}
\format{
A matrix with 3 rows and 3 columns:
\describe{
Mosquitoes in patch 1 stay with probability 0.975, move to patch 2 with probability 0.025,
mosquitoes in patch 2 and 3 stay in their patches.
}
}
\usage{
data(moveMatIndependent3)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
