% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatTaleOfTwoCities}
\alias{moveMatTaleOfTwoCities}
\title{Movement Matrix: Tale of Two Cities}
\format{
A matrix with 2 rows and 2 columns:
\describe{
Mosquitoes do not move between the two patches.
}
}
\usage{
data(moveMatTaleOfTwoCities)
}
\description{
A movement matrix for simulation with 2 patches.
}
\keyword{datasets}
