% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moveMatTriple}
\alias{moveMatTriple}
\title{Movement Matrix: Triple}
\format{
A matrix with 3 rows and 3 columns:
\describe{
All entries of matrix are 1 for testing that mosquitoes will be produced.
}
}
\usage{
data(moveMatTriple)
}
\description{
A movement matrix for simulation with 3 patches.
}
\keyword{datasets}
