% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Simulation.R
\name{multRun_Network}
\alias{multRun_Network}
\title{Run Simulation}
\usage{
multRun_Network(conM = NULL, conF = NULL, verbose = TRUE)
}
\arguments{
\item{conM}{Optional vector of \code{\link[base]{connection}} to write male population dynamics to}

\item{conF}{Optional vector of \code{\link[base]{connection}} to write female population dynamics to}

\item{verbose}{Chatty? Default is TRUE}
}
\description{
Run multiple simulations on this network
}
