% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_admPupating_deterministic_Patch}
\alias{oneDay_admPupating_deterministic_Patch}
\title{Deterministc Adult Male Pupation}
\usage{
oneDay_admPupating_deterministic_Patch()
}
\description{
Adult male emergence is calculated based on the number of male pupae multiplied by \deqn{\bigg(\overline{\xi_m} * (\theta_{e} * \theta_{p}) * (1-\mu_{ad}) * D(\theta_l,T_p) \bigg)},
where \eqn{\overline{\xi_{m}}} is the genotype-specific pupation success probability.
}
