% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_admPupating_stochastic_Patch}
\alias{oneDay_admPupating_stochastic_Patch}
\title{Stochastic Adult Male Pupation}
\usage{
oneDay_admPupating_stochastic_Patch()
}
\description{
Pupation from male pupae to adults is sampled from a binomial distribution for each genotype where the probability of pupation is given by \deqn{\bigg(\overline{\xi_m} * (\theta_{e} * \theta_{p}) * (1-\mu_{ad}) * D(\theta_l,T_p) \bigg)},
where \eqn{\overline{\xi_{m}}} is the genotype-specific pupation success probability.
}
