% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_admSurvival_deterministic_Patch}
\alias{oneDay_admSurvival_deterministic_Patch}
\title{Deterministc Adult Male Survival}
\usage{
oneDay_admSurvival_deterministic_Patch()
}
\description{
Daily adult male survival is calculated according to \deqn{\overline{\overline{Am_{[t-1]}}} * (1-\mu_{ad}) * \overline{\omega_m}}, where \eqn{\mu_{ad}} is adult mortality rate and \eqn{\overline{\omega_m}} correponds to
genotype-specific mortality effects.
}
