% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_admSurvival_stochastic_Patch}
\alias{oneDay_admSurvival_stochastic_Patch}
\title{Stochastic Adult Male Survival}
\usage{
oneDay_admSurvival_stochastic_Patch()
}
\description{
Daily adult male survival is sampled from a binomial distribution where survival probability is given by \deqn{(1-\mu_{ad}) * \overline{\omega_m}}, where \eqn{\mu_{ad}} is adult mortality rate and \eqn{\overline{\omega_m}} correponds to
genotype-specific mortality effects.
}
