% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_adoDM_stochastic_Patch}
\alias{oneDay_adoDM_stochastic_Patch}
\title{Stochastic Pupa Death and Maturation}
\usage{
oneDay_adoDM_stochastic_Patch()
}
\description{
Daily adolescent survival is sampled from a binomial distribution, where survival
probability is given by \deqn{1-\mu_{aq}}. \eqn{\mu_{aq}} corresponds
to daily non-density-dependent juvenile mortality. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
