% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_adultDeath_deterministic_Patch}
\alias{oneDay_adultDeath_deterministic_Patch}
\title{Deterministc Adult Survival}
\usage{
oneDay_adultDeath_deterministic_Patch()
}
\description{
Daily adult survival is calculated according to \deqn{\overline{\overline{Af_{[t-1]}}} * (1-\mu_{ad}) * \overline{\omega_{m/f}}},
where \eqn{\mu_{ad}} corresponds to adult mortality rate and \eqn{\overline{\omega_{m/f}}}
corresponds to genotype-specific male/female mortality effects.
}
\details{
Toxcity is also included in this adult mortality process for a specified time interval
within the simulation. Daily survival is caluclated by by \eqn{1-\mu_{tox}}. At this time,
there is no parameter for genotype-specific toxicant susceptibility

density dependent function for adults mediated by carrying capacity
function \eqn{k/{A_{t-1} + k}} applied to each genotype in the adult
population vectors (male and female density-dependence calculated
separately)
}
