% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_adultDeath_stochastic_Patch}
\alias{oneDay_adultDeath_stochastic_Patch}
\title{Stochastic Adult Survival}
\usage{
oneDay_adultDeath_stochastic_Patch()
}
\description{
Daily adult survival is sampled from a binomial distribution where survival
probability is given by \deqn{(1-\mu_{ad}) * \overline{\omega_m/f}}.
\eqn{\mu_{ad}} corresponds to adult mortality rate and \eqn{\overline{\omega_m/f}}
corresponds to genotype-specific mortality effects.
}
