% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_af1Mating_deterministic_Patch}
\alias{oneDay_af1Mating_deterministic_Patch}
\title{Deterministc Mating}
\usage{
oneDay_af1Mating_deterministic_Patch()
}
\description{
Mating is calculated as the outer product of newly emerging adult females and adult males, modulated by \eqn{\overline{\eta}}, genotype-specific male mating fitness.
}
