% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_af1Mating_stochastic_Patch}
\alias{oneDay_af1Mating_stochastic_Patch}
\title{Stochastic Mating}
\usage{
oneDay_af1Mating_stochastic_Patch()
}
\description{
Mating for each newly emerging adult female genotype is sampled from a multinomial distribution with probabilities equal to the adult male population vector multiplied by \eqn{\overline{\eta}}, genotype-specific male mating fitness.
}
