% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_af1Pupation_deterministic_Patch}
\alias{oneDay_af1Pupation_deterministic_Patch}
\title{Deterministc Adult Female Pupation}
\usage{
oneDay_af1Pupation_deterministic_Patch()
}
\description{
Adult female emergence is calculated based on the number of female pupae multiplied by \deqn{\bigg(\overline{\xi_f} * (\theta_{e} * \theta_{p}) * (1-\mu_{ad}) * D(\theta_l,T_p) \bigg)},
where \eqn{\overline{\xi_{f}}} is the genotype-specific pupation success probability.
}
