% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_af1Survival_deterministic_Patch}
\alias{oneDay_af1Survival_deterministic_Patch}
\title{Deterministc Adult Female Survival}
\usage{
oneDay_af1Survival_deterministic_Patch()
}
\description{
Daily adult female survival is calculated according to \deqn{\overline{\overline{Af_{[t-1]}}} * (1-\mu_{ad}) * \overline{\omega_f}}, where \eqn{\mu_{ad}} is adult mortality rate and \eqn{\overline{\omega_f}} correponds to
genotype-specific mortality effects.
}
