% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_af1Survival_stochastic_Patch}
\alias{oneDay_af1Survival_stochastic_Patch}
\title{Stochastic Adult Female Survival}
\usage{
oneDay_af1Survival_stochastic_Patch()
}
\description{
Daily adult female survival is sampled from a binomial distribution where survival probability is given by \deqn{(1-\mu_{ad}) * \overline{\omega_f}}, where \eqn{\mu_{ad}} is adult mortality rate and \eqn{\overline{\omega_f}} corresponds to
genotype-specific mortality effects.
}
