% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_calcCumulativeLarvalDensityDependentFactor_Patch}
\alias{oneDay_calcCumulativeLarvalDensityDependentFactor_Patch}
\title{Calculate Cumulative Larval density-dependent Mortality}
\usage{
oneDay_calcCumulativeLarvalDensityDependentFactor_Patch()
}
\description{
Calculate \deqn{F(L[t])=\Bigg(\frac{\alpha}{\alpha+\sum{\overline{L[t]}}}\Bigg)^{1/T_l}\\}, the cumulative effect of density-dependence on larval mortality over the entire duration of larval stage.
}
