% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_calcCumulativePupaDensityDependentFactor_Patch}
\alias{oneDay_calcCumulativePupaDensityDependentFactor_Patch}
\title{Calculate Cumulative Pupae density-dependent Mortality}
\usage{
oneDay_calcCumulativePupaDensityDependentFactor_Patch()
}
\description{
Calculate \deqn{D(\theta_l,T_{P}) = \left\{ \begin{array}{ll} \theta_{l[0]}^{'}=\theta_l 								& \quad i = 0 \\ \theta_{l[i+1]}^{'} = \theta_{l[i]}^{'} *F(\overline{L_{[t-i-T_P]}})	& \quad i \leq T_{P} \end{array} \right.}, the cumulative effect of density-dependence on larval mortality over the entire duration of larval and pupal stages.
}
