% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_calcLarvalDensityDependentFactor_Patch}
\alias{oneDay_calcLarvalDensityDependentFactor_Patch}
\title{Calculate Larval density-dependent Mortality}
\usage{
oneDay_calcLarvalDensityDependentFactor_Patch()
}
\description{
Calculate \deqn{D(\theta_l,T_x) = \left\{ \begin{array}{ll} \theta_{l[0]}^{'}=\theta_l	& \quad i = 0 \\ \theta_{l[i+1]}^{'} = \theta_{l[i]}^{'} *F(\overline{L_{[t-i-T_x]}})	& \quad i \leq T_l \end{array}    \right.},
the effect of density-dependence on larval mortality for one day.
}
