% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_eggDM_deterministic_Patch}
\alias{oneDay_eggDM_deterministic_Patch}
\title{Deterministic Egg Death and Pupation}
\usage{
oneDay_eggDM_deterministic_Patch()
}
\description{
Daily egg survival is calculated according to \deqn{\overline{E_{[t-1]}} * (1-\mu_{aq})},
where \eqn{\mu_{aq}} corresponds to daily non-density-dependent aquatic mortality.
Eggs transition into larvae at the end of \eqn{T_e}. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
