% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Releases.R
\name{oneDay_eggReleases2adults_Patch}
\alias{oneDay_eggReleases2adults_Patch}
\title{Release Eggs into Larval Window}
\usage{
oneDay_eggReleases2adults_Patch()
}
\description{
Get eggs that were released into larval window t - aquaStageDuration days
ago and put them into the oviposit vector
}
