% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Releases.R
\name{oneDay_eggReleases2eggs_Patch}
\alias{oneDay_eggReleases2eggs_Patch}
\title{Release Eggs into hatching eggs}
\usage{
oneDay_eggReleases2eggs_Patch()
}
\description{
This function performs egg releases into hatching eggs. This allows them
to under-go density-dependence
}
