% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_eggReleases_Patch}
\alias{oneDay_eggReleases_Patch}
\title{Release Eggs in a Patch}
\usage{
oneDay_eggReleases_Patch()
}
\description{
Based on this patch's release schedule, \code{\link{generateReleaseVector}},
this function handles daily egg releases.
}
