% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_eggsFract2_stochastic_Patch}
\alias{oneDay_eggsFract2_stochastic_Patch}
\title{Stochastic Larval Pupation}
\usage{
oneDay_eggsFract2_stochastic_Patch()
}
\description{
Calculate the number of larvae that will pupate prior to calculating density-dependent effects in \code{\link{oneDay_calcCumulativeLarvalDensityDependentFactor_Patch}}.
}
