% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Releases.R
\name{oneDay_femaleReleases_Patch}
\alias{oneDay_femaleReleases_Patch}
\title{Release Female Mosquitoes in a Patch}
\usage{
oneDay_femaleReleases_Patch()
}
\description{
Based on this patch's release schedule, handle daily releases.
}
