% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_hatchingFract_deterministic_Patch}
\alias{oneDay_hatchingFract_deterministic_Patch}
\title{Deterministc Fraction of Eggs Maturing to Hatch}
\usage{
oneDay_hatchingFract_deterministic_Patch()
}
\description{
Calculate the fraction of hatching eggs accounting for delay given by: \eqn{\overline{O(T_e)}}
}
