% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_hatchingFract_stochastic_Patch}
\alias{oneDay_hatchingFract_stochastic_Patch}
\title{Stochastic Fraction of Eggs Maturing to Hatch}
\usage{
oneDay_hatchingFract_stochastic_Patch()
}
\description{
Calculate the fraction of hatching eggs accounting for delay given by: \eqn{\overline{O(T_e)}}
Stochasticity is introduced by assuming \eqn{\overline{O(T_e)}} defines the mean of a Poisson distributed number of eggs that can hatch.
}
