% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{oneDay_initOutput_Patch}
\alias{oneDay_initOutput_Patch}
\title{Initialize Output from Focal Patch}
\usage{
oneDay_initOutput_Patch()
}
\description{
Writes output to the text connections specified in the enclosing \code{\link{Network}}.
}
