% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larHatching_deterministic_Patch}
\alias{oneDay_larHatching_deterministic_Patch}
\title{Stochastic Egg Hatching to Larval Stage}
\usage{
oneDay_larHatching_deterministic_Patch()
}
\description{
Calculate the eggs that have survived and hatched during a day given by: \eqn{\overline{O(T_e)}* \theta_{e}}
}
