% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larPupating_deterministic_Patch}
\alias{oneDay_larPupating_deterministic_Patch}
\title{Deterministc Larval Pupation}
\usage{
oneDay_larPupating_deterministic_Patch()
}
\description{
Calculate the number of larvae that have transformed into pupae given by \eqn{\overline{O(T_e+T_l)} * \theta_{e} * D(\theta_l,0)}
}
