% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larSurvival_deterministic_Patch}
\alias{oneDay_larSurvival_deterministic_Patch}
\title{Deterministc Larval Survival}
\usage{
oneDay_larSurvival_deterministic_Patch()
}
\description{
Calculate the number of larvae surviving from day to day, given by:
\deqn{\overline{L_{[t-1]}} * (1-\mu_{l}) * F(\overline{L_{[t-1]})}}
}
