% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larSurvival_stochastic_Patch}
\alias{oneDay_larSurvival_stochastic_Patch}
\title{Stochastic Larval Survival}
\usage{
oneDay_larSurvival_stochastic_Patch()
}
\description{
Calculate the number of larvae surviving from day to day, given by:
\deqn{\overline{L_{[t-1]}} * (1-\mu_{l}) * F(\overline{L_{[t-1]})}}
Stochasticity is introduced by assuming \eqn{(1-\mu_{l}) * F(\overline{L_{[t-1]})}} defines binomial likelihood of survival for each
genotype of larvae.
}
