% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larvaDM_deterministic_Patch}
\alias{oneDay_larvaDM_deterministic_Patch}
\title{Deterministic Larva Death and Pupation}
\usage{
oneDay_larvaDM_deterministic_Patch()
}
\description{
Calculate the number of larvae surviving from day to day, given by:
\deqn{\overline{L_{[t-1]}} * (1-\mu_{aq}) * F(L)}. F(L), the density dependence
is calculated as \deqn{F(L[t])=\Bigg(\frac{\alpha}{\alpha+\sum{\overline{L[t]}}}\Bigg)^{1/T_l}}.
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
Pupation has no parameters, so the final day of larvae naturally enter the pupal state.
}
