% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_larvaDM_stochastic_Patch}
\alias{oneDay_larvaDM_stochastic_Patch}
\title{Stochastic Larva Death and Pupation}
\usage{
oneDay_larvaDM_stochastic_Patch()
}
\description{
The daily number of larvae surviving is drawn from a binomial distribution, where
survival probability is given by \deqn{(1-\mu_{aq}) * F(L)}. F(L), the density dependence
is calculated as \deqn{F(L[t])=\Bigg(\frac{\alpha}{\alpha+\sum{\overline{L[t]}}}\Bigg)^{1/T_l}}.
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
Pupation has no parameters, so the final day of larvae naturally enter the pupal state.
}
