% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Releases.R
\name{oneDay_maleReleases_Patch}
\alias{oneDay_maleReleases_Patch}
\title{Release Male Mosquitoes in a Patch}
\usage{
oneDay_maleReleases_Patch()
}
\description{
Based on this patch's release schedule, handle daily releases.
}
