% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_mating_stochastic_Patch}
\alias{oneDay_mating_stochastic_Patch}
\title{Stochastic Mating}
\usage{
oneDay_mating_stochastic_Patch()
}
\description{
Mating for each newly emerging adult female genotype is sampled from a multinomial
distribution with probabilities equal to the adult male population vector
multiplied by \eqn{\overline{\overline{\eta}}}, the genotype-specific
male mating fitness. \eqn{\overline{\overline{\eta}}} corresponds a number of females calculated using a
binomial distribution, with the probability of mating for each femal equal to \eqn{litters/365} or the probability
of a female having a litter on any given day (rows)
and male (columns) genotypes, to perform any type of assortative mating. \cr
}
