% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_numMaleFemale_deterministic_Patch}
\alias{oneDay_numMaleFemale_deterministic_Patch}
\title{Deterministc Sex Ratio}
\usage{
oneDay_numMaleFemale_deterministic_Patch()
}
\description{
Calculate the number of males, \eqn{(1-\overline{\phi}) *  \overline{E^{'}}} and females, \eqn{\overline{\phi} *  \overline{E^{'}}}
}
