% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_numMaleFemale_stochastic_Patch}
\alias{oneDay_numMaleFemale_stochastic_Patch}
\title{Stochastic Sex Ratio}
\usage{
oneDay_numMaleFemale_stochastic_Patch()
}
\description{
Calculate the number of males, \eqn{(1-\overline{\phi}) *  \overline{E^{'}}} and females, \eqn{\overline{\phi} *  \overline{E^{'}}}
These counts are follow a binomial distribution.
}
