% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_nursingDM_deterministic_Patch}
\alias{oneDay_nursingDM_deterministic_Patch}
\title{Deterministic Larva Death and Maturation}
\usage{
oneDay_nursingDM_deterministic_Patch()
}
\description{

}
