% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_nursingDM_stochastic_Patch}
\alias{oneDay_nursingDM_stochastic_Patch}
\title{Stochastic Larva Death and Maturation}
\usage{
oneDay_nursingDM_stochastic_Patch()
}
\description{
The daily number of nursing pups surviving is drawn from a binomial distribution, where
survival probability is given by \deqn{1-\mu_{aq}}
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
Maturation has no parameters, so the final day of nursing pups naturally enter the adolescent state.
}
