% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_oviposit_deterministic_Patch}
\alias{oneDay_oviposit_deterministic_Patch}
\title{Deterministic Oviposition}
\usage{
oneDay_oviposit_deterministic_Patch()
}
\description{
Calculate the number of eggs oviposited by female mosquitoes following:
\deqn{\overline{O(T_x)} = \sum_{j=1}^{n} \Bigg( \bigg( (\beta*\overline{s} * \overline{ \overline{Af_{[t]}}}) * \overline{\overline{\overline{Ih}}} \bigg) * \Lambda \Bigg)^{\top}_{ij}}
}
