% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_pupaDM_deterministic_Patch}
\alias{oneDay_pupaDM_deterministic_Patch}
\title{Deterministic Pupa Death and Pupation}
\usage{
oneDay_pupaDM_deterministic_Patch()
}
\description{
Daily pupa survival is calculated according to \deqn{\overline{P_{[t-1]}} * (1-\mu_{aq})},
where \eqn{\mu_{aq}} corresponds to daily non-density-dependent aquatic mortality. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
