% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_pupaDM_stochastic_Patch}
\alias{oneDay_pupaDM_stochastic_Patch}
\title{Stochastic Pupa Death and Pupation}
\usage{
oneDay_pupaDM_stochastic_Patch()
}
\description{
Daily pupa survival is sampled from a binomial distribution, where survival
probability is given by \deqn{1-\mu_{aq}}. \eqn{\mu_{aq}} corresponds
to daily non-density-dependent aquatic mortality. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
