% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_pupation_deterministic_Patch}
\alias{oneDay_pupation_deterministic_Patch}
\title{Deterministic Pupation}
\usage{
oneDay_pupation_deterministic_Patch()
}
\description{
Pupa first undergo one extra day of survival, calculated as \deqn{\overline{P_{[t-1]}} * (1-\mu_{ad})}.
This is an artifact of the conversion from continuous to discrete time (as mentioned
in the original Hancock paper this model is derived from). \cr
Then, pupation into adult males is calculated as \deqn{(1-\overline{\phi}) * \overline{P_{[t]}}}
and into adult females as \deqn{\overline{\phi} * \overline{P_{[t]}}}
}
