% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_pupation_stochastic_Patch}
\alias{oneDay_pupation_stochastic_Patch}
\title{Stochastic Pupation}
\usage{
oneDay_pupation_stochastic_Patch()
}
\description{
Pupa first undergo one extra day of survival, calculated as a binomial over
\deqn{\overline{P_{[t-1]}} * (1-\mu_{ad})}.
This is an artifact of the conversion from continuous to discrete time (as mentioned
in the original Hancock paper this model is derived from). \cr
Then, pupation is sampled from a binomial, where \eqn{(1-\overline{\phi})} is
the genotype-specific probability of becoming male, and \eqn{\overline{\phi}}
is the genotype-specific of becoming female.
}
