% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_releases_Patch}
\alias{oneDay_releases_Patch}
\title{Release Male/Female/Mated-Female Mosquitoes in a Patch}
\usage{
oneDay_releases_Patch()
}
\description{
Based on this patch's release schedule, \code{\link{generateReleaseVector}},
this function handles daily releases.
}
