% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_updatePopulation_Patch}
\alias{oneDay_updatePopulation_Patch}
\title{Shift Population}
\usage{
oneDay_updatePopulation_Patch()
}
\description{
Update larval and adult populations daily at end of time step, calls \code{\link{shiftAndUpdatePopVector}}
}
