% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{oneDay_writeOutput_Patch}
\alias{oneDay_writeOutput_Patch}
\title{Write Output from Focal Patch}
\usage{
oneDay_writeOutput_Patch()
}
\description{
Writes output to the text connections specified in the enclosing \code{\link{Network}}.
}
