% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{parameterizeMGDrivE}
\alias{parameterizeMGDrivE}
\title{parameterizeMGDrivE}
\usage{
parameterizeMGDrivE(
  runID = 1L,
  nPatch,
  simTime,
  sampTime = 1L,
  moveVar = 1000L,
  tGest = 19L,
  tNursing = 23L,
  tAdo = 37L,
  beta = 6,
  litters = 7.5,
  muAd = I(1/690),
  k,
  theta = 22.4,
  AdPopRatio_F,
  AdPopRatio_M,
  inheritanceCube
)
}
\arguments{
\item{runID}{Begin counting runs with this set of parameters from this value}

\item{nPatch}{Number of \code{\link{Patch}}}

\item{simTime}{Maximum time to run simulation}

\item{sampTime}{Times to sample, used as tNow \%\% sampTime, default is every day}

\item{moveVar}{Variance of stochastic movement (not used in diffusion model of migration).
It affects the concentration of probability in the Dirchlet simplex, small values
lead to high variance and large values lead to low variance.}

\item{tGest}{Length of gestation stage}

\item{tNursing}{Length of nursing stage}

\item{tAdo}{Length of adolescent stage}

\item{beta}{Female litter size of wild-type}

\item{litters}{average number of litters per female mouse per year}

\item{muAd}{Wild-type daily adult mortality (1/muAd is average wild-type lifespan)}

\item{k}{Single number or vector of adult population carrying capacity at equilibrium
(single number implies all patches have the same population, but populations for
individual patches can be specified with a vector)}

\item{theta}{Shape parameter for mediating the carrying capacity for adult mice}

\item{AdPopRatio_F}{May be empty; if not, a vector gives the wildtype gene frequencies
among adult females at the beginning of simulation or a matrix provides different
initial frequencies for each patch (every row is a different patch, must have nrow = nPatch)}

\item{AdPopRatio_M}{May be empty; if not, a vector gives the wildtype gene frequencies
among adult males at the beginning of simulation or a matrix provides different
initial frequencies for each patch (every row is a different patch, must have nrow = nPatch)}

\item{inheritanceCube}{Inheritance cube to check/set population ratios at the beginning of the simulation}
}
\description{
Generate parameters for simulation on a \code{\link{Network}}.
Parameters include: average generation time \eqn{g}, population growth rate \eqn{R_{m}},
juvenile mortality \eqn{\mu_{Ju}}, and juvenile survival \eqn{\theta_{Ju}}, which
are shared between patches and calculated by \code{\link{calcAverageGenerationTime}},
\code{\link{calcPopulationGrowthRate}}, and \code{\link{calcLarvalStageMortalityRate}}. \cr
Patch-specific parameters \eqn{\alpha} and \eqn{L_{eq}}
are calculated for each patch by \code{\link{calcDensityDependentDeathRate}}
and \code{\link{calcLarvalPopEquilibrium}}.
}
\examples{
# using default parameters for 2 patches
#  using different population sizes for patches
simPars <- parameterizeMGDrivE(nPatch = 2, simTime = 365,
                               k = c(100,200), inheritanceCube = cubeMendelian())

}
