% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{primePopMatrixArray}
\alias{primePopMatrixArray}
\title{Create a primed population array of matrices}
\usage{
primePopMatrixArray(primingMatrix, memoryWindow)
}
\arguments{
\item{primingMatrix}{a named matrix population}

\item{memoryWindow}{integer size of list structure}
}
\description{
Primes an array for the population history to be stored. The length of the
array is equal to the window required for the model to run
(in our specific case it is equal to the sum of aquatic stages lengths).
}
