% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantileC}
\alias{quantileC}
\title{Quantiles Function}
\usage{
quantileC(Trials, Probs)
}
\arguments{
\item{Trials}{Integer matrix to calculate quantiles over}

\item{Probs}{Vector of quantiles}
}
\value{
Numeric Matrix
}
\description{
Calculate the given quantiles of a matrix.
}
\details{
This function calculates the given quantiles over the rows of an
integer matrix. It uses method 8 of the stat::quantiles() function. It gives
the same result, to numerical accuracy, and is designed to handle matrix input.
It is only designed to work on integer matrices!
}
