% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Dirichlet Distribution}
\usage{
rDirichlet(migrationPoint)
}
\arguments{
\item{migrationPoint}{Vector of weights for draws. Must be positive.}
}
\description{
Make a single draw from a Dirichlet distribution with the shape parameter
one.
}
