% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Simulation.R
\name{reset_Network}
\alias{reset_Network}
\title{Reset Network}
\usage{
reset_Network(verbose = TRUE)
}
\arguments{
\item{verbose}{Chatty? Default is TRUE}
}
\description{
Reset a \code{\link{Network}} between runs, useful for Monte Carlo simulation.
This calls \code{\link{reset_Patch}} on each patch
and resets \code{tNow = 2} and increments the \code{runID}.
}
