% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{reset_Patch}
\alias{reset_Patch}
\title{Reset Patch to Initial Conditions}
\usage{
reset_Patch()
}
\description{
Resets a patch to its initial configuration so that a new one does not have to be created and
allocated in the network (for Monte Carlo simulation).
}
