% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{retrieveOutput}
\alias{retrieveOutput}
\title{Retrieve Output}
\usage{
retrieveOutput(directory, genotypes)
}
\arguments{
\item{directory}{directory where output was written to; must not end in path seperator}

\item{genotypes}{character vector of possible genotypes; found in \code{driveCube$genotypesID}}
}
\description{
Read in output from directory. The resulting object will be a nested list;
outermost nesting dimension indexes runID, within runID elements are split by sex
and innermost nesting is over patches.
}
