% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{set_ADMnew_Patch}
\alias{set_ADMnew_Patch}
\title{Set ADMnew}
\usage{
set_ADMnew_Patch(count, genotype_M)
}
\arguments{
\item{count}{number of mosquitoes}

\item{genotype_M}{genotype of male}
}
\description{
Set an element in new ADM males
}
