% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{set_AF1new_Patch}
\alias{set_AF1new_Patch}
\title{Set AF1new}
\usage{
set_AF1new_Patch(count, genotype_F, genotype_M)
}
\arguments{
\item{count}{number of mosquitoes}

\item{genotype_F}{genotype of female (row of matrix)}

\item{genotype_M}{genotype of male (column of matrix)}
}
\description{
Set an element in new AF1 females
}
