% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{set_migrationFemale_Network}
\alias{set_migrationFemale_Network}
\title{Set Female Migration Matrix}
\usage{
set_migrationFemale_Network(migrationFemale)
}
\arguments{
\item{migrationFemale}{matrix object (rows must sum to one)}
}
\description{
Sets \code{link{private$migrationFemale}} field
}
