% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{set_migrationMale_Network}
\alias{set_migrationMale_Network}
\title{Set Male Migration Matrix}
\usage{
set_migrationMale_Network(migrationMale)
}
\arguments{
\item{migrationMale}{matrix object (rows must sum to one)}
}
\description{
Sets \code{link{private$migrationMale}} field
}
