% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\name{set_NetworkPointer_Patch}
\alias{set_NetworkPointer_Patch}
\title{Set Network Pointer}
\usage{
set_NetworkPointer_Patch(NetworkPointer)
}
\arguments{
\item{NetworkPointer}{A \code{\link{Network}} object}
}
\description{
Set a reference to the enclosing \code{\link{Network}} object
}
