% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{set_population_deterministic_Patch}
\alias{set_population_deterministic_Patch}
\title{Set Initial Population Deterministic}
\usage{
set_population_deterministic_Patch(
  adultEQ = adultEQ,
  larvalEQ = larvalEQ,
  adultRatioF = adultRatioF,
  adultRatioM = adultRatioM,
  larvalRatio = larvalRatio,
  timeAq = timeAq,
  muAq = muAq,
  alpha = alpha
)
}
\arguments{
\item{adultEQ}{Equilibrium number of adults}

\item{larvalEQ}{Equilibrium number of larvae}

\item{adultRatioF}{Genotype specific ratio for adult females}

\item{adultRatioM}{Genotype specific ratio for adult males}

\item{larvalRatio}{Genotype specific ratio for larvae}

\item{timeAq}{Time for each aquatic stage}

\item{muAq}{Aquatic death rate}

\item{alpha}{Density-dependent centering parameter}
}
\description{
Calls \code{\link{set_initialPopulation_Patch}} to initialize a steady-state
population distribution.
}
