% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Setup.R
\name{setupMGDrivE}
\alias{setupMGDrivE}
\title{Setup MGDrivE}
\usage{
setupMGDrivE(stochasticityON = FALSE, verbose = TRUE)
}
\arguments{
\item{stochasticityON}{Enable/disable stochastic simulation. Default is FALSE, implying deterministic simulation}

\item{verbose}{Chatty? Default is TRUE}
}
\description{
Initialize methods in \code{\link{Patch}} to run deterministic or stochastic simulations.
This sets internal function definitions so that \code{\link{oneRun_Network}}
and \code{\link{multRun_Network}} run either deterministic or stochastic functions.
}
\examples{
# run deterministic MGDrivE
setupMGDrivE(stochasticityON = FALSE)

# run stochastic MGDrivE
setupMGDrivE(stochasticityON = TRUE)

}
