% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shiftAndUpdatePopVector}
\alias{shiftAndUpdatePopVector}
\title{Shift a Vector}
\usage{
shiftAndUpdatePopVector(popVector, newPop)
}
\arguments{
\item{popVector}{List of population vectors of length(Tegg+Tlarva+Tpupa)}

\item{newPop}{Vector of length equal to the number of genotypes}
}
\description{
Shift a population vector by one day and insert the new population.
This was written to remove the dependency "binhf".
}
