% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SymCubeC}
\alias{SymCubeC}
\title{Make a Symmetric Cube}
\usage{
SymCubeC(lowerMat)
}
\arguments{
\item{lowerMat}{A lower-triangular matrix of depth 1 or more}
}
\description{
This function makes a lower-triangular cube symmetric over the z-axis.
It was written to remove the dependency "Matrix".
It is only used in building cubes.
}
