% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Setup.R
\name{turnStochasticityOnOrOff}
\alias{turnStochasticityOnOrOff}
\title{Enable or Disable Stochastic Model}
\usage{
turnStochasticityOnOrOff(on = TRUE)
}
\arguments{
\item{on}{enable/disable stochastic behaviour}
}
\description{
Set switches for deterministic or stochastic model
}
