% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{aggregateFemales}
\alias{aggregateFemales}
\title{Aggregate Female Output by Genotype}
\usage{
aggregateFemales(directory, genotypes, remove = FALSE, multiCore = FALSE)
}
\arguments{
\item{directory}{Directory where output was written to; must not end in path seperator}

\item{genotypes}{Character vector of possible genotypes; found in \code{driveCube$genotypesID}}

\item{remove}{Boolean flag to remove original (unaggregated) file}

\item{multiCore}{Write output using multiple cores? Default is FALSE}
}
\description{
Aggregate over male mate genotype to convert female matrix output into vector output.
}
