% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcAverageGenerationTime}
\alias{calcAverageGenerationTime}
\title{Calculate Average Generation Time}
\usage{
calcAverageGenerationTime(stagesDuration, tAd)
}
\arguments{
\item{stagesDuration}{Vector of lengths of aquatic stages, \eqn{T_{e}, T_{l}, T_{p}}}

\item{tAd}{Vector of lengths of aquatic stages, \eqn{T_{a}}}
}
\description{
Calculate \eqn{g}, average generation time, given by: \deqn{g=T_e+T_l+T_p+\frac{1}{\mu_{ad}}}
}
