% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcGammaKernel}
\alias{calcGammaKernel}
\title{Calculate Gamma Stochastic Matrix}
\usage{
calcGammaKernel(distMat, shape, rate)
}
\arguments{
\item{distMat}{Distance matrix from \code{\link[MGDrivE]{calcVinEll}}}

\item{shape}{Shape parameter of \code{\link[stats]{GammaDist}} distribution}

\item{rate}{Rate parameter of \code{\link[stats]{GammaDist}} distribution}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calcVinEll}}, calculate a
stochastic matrix where one step movement probabilities follow a gamma density.
}
\details{
The distribution and density functions for the gamma kernel are given below:
\deqn{
F(x)=\frac{1}{\Gamma(\alpha)}\gamma(\alpha,\beta x)
}
\deqn{
f(x)=\frac{\beta^{\alpha}}{\Gamma(\alpha)}x^{\alpha-1}e^{-\beta x}
}
where \eqn{\Gamma(\alpha)} is the Gamma function, \eqn{\gamma(\alpha,\beta x)} is the lower incomplete
gamma function, and \eqn{\alpha,\beta} are the shape and rate parameters, respectively.
}
\examples{
# setup distance matrix
# two-column matrix with latitude/longitude, in degrees
latLong = cbind(runif(n = 5, min = 0, max = 90),
                runif(n = 5, min = 0, max = 180))

# Vincenty Ellipsoid  distance formula
distMat = calcVinEll(latLongs = latLong)

# calculate gamma distribution over distances
#  shape and rate are just for example
kernMat = calcGammaKernel(distMat = distMat, shape = 1, rate = 1)

}
