% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcLarvalPopEquilibrium}
\alias{calcLarvalPopEquilibrium}
\title{Calculate Equilibrium Larval Population}
\usage{
calcLarvalPopEquilibrium(alpha, Rm)
}
\arguments{
\item{alpha}{See \code{\link{calcDensityDependentDeathRate}}}

\item{Rm}{See \code{\link{calcPopulationGrowthRate}}}
}
\description{
Equilibrium larval population size to sustain adult population.
}
