% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcLognormalKernel}
\alias{calcLognormalKernel}
\title{Calculate Lognormal Stochastic Matrix}
\usage{
calcLognormalKernel(distMat, meanlog, sdlog)
}
\arguments{
\item{distMat}{Distance matrix from \code{\link[MGDrivE]{calcVinEll}}}

\item{meanlog}{Log mean of \code{\link[stats]{Lognormal}} distribution}

\item{sdlog}{Log standard deviation of \code{\link[stats]{Lognormal}} distribution}
}
\description{
Given a distance matrix from \code{\link[MGDrivE]{calcVinEll}},
calculate a stochastic matrix where one step movement probabilities follow a lognormal density.
}
\details{
The distribution and density functions for the lognormal kernel are given below:
\deqn{
F(x)=\frac{1}{2} + \frac{1}{2} \mathrm{erf}[\frac{\mathrm{ln}x-\mu}{\sqrt{2}\sigma}]
}
\deqn{
f(x)=\frac{1}{x\sigma\sqrt{2\pi}}\mathrm{exp}\left( -\frac{(\mathrm{ln}x-\mu)^{2}}{2\sigma^{2}} \right)
}
where \eqn{\mu} is the mean on the log scale, and \eqn{\sigma} is the standard deviation on the log scale.
}
\examples{
# setup distance matrix
# two-column matrix with latitude/longitude, in degrees
latLong = cbind(runif(n = 5, min = 0, max = 90),
                runif(n = 5, min = 0, max = 180))

# Vincenty Ellipsoid  distance formula
distMat = calcVinEll(latLongs = latLong)

# calculate lognormal distribution over distances
#  mean and standard deviation are just for example
kernMat = calcLognormalKernel(distMat = distMat, meanlog = 100, sdlog = 10)

}
