% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Parameters.R
\name{calcMemoryWindow}
\alias{calcMemoryWindow}
\title{Calculate Memory Window}
\usage{
calcMemoryWindow(stagesDuration)
}
\arguments{
\item{egg}{length of egg stage (days)}

\item{larva}{length of larval stage (days)}

\item{pupa}{length of pupal stage (days)}
}
\description{
Calculates the necessary window of population history required for the model to work
}
