% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cube-Auxiliary.R
\name{cube2csv}
\alias{cube2csv}
\title{Export a Cube to .csv}
\usage{
cube2csv(cube, directory, digits = 3)
}
\arguments{
\item{cube}{a cube object (see \code{\link{MGDrivE-Cube}} for options)}

\item{directory}{directory to write .csv files to}

\item{digits}{number of significant digits to retain in .csv output}
}
\description{
Export a cube as multiple .csv files (one for each genotype; slices of z-axis).
}
