% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Class.R
\name{get_patchReleases_Network}
\alias{get_patchReleases_Network}
\title{Get Patch Release Schedule}
\usage{
get_patchReleases_Network(ix, sex = "M")
}
\arguments{
\item{ix}{index of patch}

\item{sex}{character in 'M', 'F'}
}
\description{
Return the release schedule for a patch for male or female
}
