% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Simulation.R
\name{multRun_Network}
\alias{multRun_Network}
\title{Run Simulation}
\usage{
multRun_Network(verbose = TRUE)
}
\arguments{
\item{verbose}{Chatty? Default is TRUE}
}
\description{
Run multiple simulations on this network
}
