% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_adultDeath_deterministic_Patch}
\alias{oneDay_adultDeath_deterministic_Patch}
\title{Deterministic Adult Survival}
\usage{
oneDay_adultDeath_deterministic_Patch()
}
\description{
Daily adult survival is calculated according to \deqn{\overline{\overline{Af_{[t-1]}}} * (1-\mu_{ad}) * \overline{\omega_{m/f}}},
where \eqn{\mu_{ad}} corresponds to adult mortality rate and \eqn{\overline{\omega_{m/f}}}
corresponds to genotype-specific male/female mortality effects.
}
