% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_gestDM_stochastic_Patch}
\alias{oneDay_gestDM_stochastic_Patch}
\title{Stochastic Gestating Pup Death and Maturation}
\usage{
oneDay_gestDM_stochastic_Patch()
}
\description{
Daily gestating pup survival is sampled from a binomial distribution, where survival
probability is given by \eqn{1-\mu_{aq}}. \eqn{\mu_{aq}} corresdponds
to daily non-density-dependent juvenile mortality. \cr
Gestating Pups transition into nursing pups at the end of \eqn{T_e}. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
