% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_mating_deterministic_Patch}
\alias{oneDay_mating_deterministic_Patch}
\title{Deterministic Mating}
\usage{
oneDay_mating_deterministic_Patch()
}
\description{
Mating is calculated as the outer product of newly emerging adult females and
all-current adult males, modulated by \eqn{\overline{\overline{\eta}}}, the genotype-specific
male mating fitness. \eqn{\overline{\overline{\eta}}} corresponds to all female (rows)
and male (columns) genotypes, to perform any type of assortative mating. \cr
If there are no adult males, the unmated females experience one day of death,
calculated as \deqn{\overline{Af_t} * (1-\mu_{ad}) * \overline{\omega_f}}, and
remain unmated until tomorrow.
}
