% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_mating_stochastic_Patch}
\alias{oneDay_mating_stochastic_Patch}
\title{Stochastic Mating}
\usage{
oneDay_mating_stochastic_Patch()
}
\description{
Mating for each newly emerging adult female genotype is sampled from a multinomial
distribution with probabilities equal to the adult male population vector
multiplied by \eqn{\overline{\overline{\eta}}}, the genotype-specific
male mating fitness. \eqn{\overline{\overline{\eta}}} corresponds to all female (rows)
and male (columns) genotypes, to perform any type of assortative mating. \cr
If there are no adult males, the unmated females experience one day of death,
sampled from a binomial distribution parameterized by \deqn{(1-\mu_{ad}) * \overline{\omega_f}},
and remain unmated until tomorrow.
}
