% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Migration.R
\name{oneDay_migrationIn_Patch}
\alias{oneDay_migrationIn_Patch}
\title{Inbound Migration}
\usage{
oneDay_migrationIn_Patch(maleIn, femaleIn)
}
\arguments{
\item{maleIn}{Vector of inbound migration}

\item{femaleIn}{Matrix of inbound migration}
}
\description{
Accumulate all inbound migration to this patch.
}
