% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Migration.R
\name{oneDay_migrationOut_deterministic_Patch}
\alias{oneDay_migrationOut_deterministic_Patch}
\title{Deterministic Oubound Migration from a Patch}
\usage{
oneDay_migrationOut_deterministic_Patch()
}
\description{
Deterministc model of outbound migration of \code{AF1new} females from this patch, fills up the \code{femaleMigration} array.
}
