% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Simulation.R
\name{oneRun_Network}
\alias{oneRun_Network}
\title{Run Simulation}
\usage{
oneRun_Network(conADM = NULL, conAF1 = NULL, verbose = TRUE)
}
\arguments{
\item{conADM}{Optional \code{\link[base]{connection}} to write male population dynamics to}

\item{conAF1}{Optional \code{\link[base]{connection}} to write female population dynamics to}

\item{verbose}{Chatty? Default is TRUE}
}
\description{
Run a single simulation on this network.
}
