% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network-Simulation.R
\name{oneRun_Network}
\alias{oneRun_Network}
\title{Run Simulation}
\usage{
oneRun_Network(conADM = NULL, conAF1 = NULL)
}
\arguments{
\item{conADM}{an optional \code{\link[base]{connection}} to write male population dynamics to, if \code{NULL} use the directory specified in the constructor of \code{\link{Network}} with the current runID appended to the file.}

\item{conAF1}{an optional \code{\link[base]{connection}} to write female population dynamics to, if \code{NULL} use the directory specified in the constructor of \code{\link{Network}} with the current runID appended to the file.}
}
\description{
Run a single simulation on this network.
}
