% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Graphics.R
\name{plotMGDrivEMult}
\alias{plotMGDrivEMult}
\title{Plot}
\usage{
plotMGDrivEMult(readDir, whichPatches = NULL, totalPop = FALSE,
                       nonZeroGen = FALSE, lwd = 0.75, alpha = 0.75)
}
\arguments{
\item{readDir}{Directory to find repetition folders in}

\item{whichPatches}{Vector of patches to plot, must be less than 15. Default is NULL if less than 15 patches}

\item{totalPop}{Boolean, to plot the total population or not. Default is FALSE}

\item{nonZeroGen}{Boolean, to plot genotypes that are always zero in simulation}

\item{lwd}{Double, specify the line width for plotting}

\item{alpha}{Double, specify the opacity for plotting}
}
\description{
Plots several traces from MGDrive, assuming each set is another repetition
from the same experiment. \cr
Given the readDirectory, this function assumes the follow file structure: \cr
\itemize{
\item readDirectory
\itemize{
\item repetition 1
\itemize{
\item patch 1
\item patch 2
\item patch 3
}
\item repetition 2
\itemize{
\item patch 1
\item patch 2
\item patch 3
}
\item repetition 3
\item repetition 4
\item ...
}
}
}
\details{
This function plots output from one run or one set of runs after being analyzed. Setting totalPop
to FALSE keeps it from plotting the total population. NonZeroGen accounts for
genotypes that could exist, but are not created in the simulation. Default
is FALSE, as this is easier to read on a plot.
}
\examples{
# Requires the user to have run MGDrivE with repetitions and keep them in the
#  directory organization shown above.
# See vignette for complete example

}
