% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{set_population_stochastic_Patch}
\alias{set_population_stochastic_Patch}
\title{Set Initial Population Stochastic}
\usage{
set_population_stochastic_Patch(
  k = k,
  adultRatioF = adultRatioF,
  adultRatioM = adultRatioM,
  muAd = muAd,
  timeAd = timeAd
)
}
\arguments{
\item{k}{Carrying capacity constant}

\item{adultRatioF}{Genotype specific ratio for adult females}

\item{adultRatioM}{Genotype specific ratio for adult males}

\item{muAd}{daily survival of adult stage}

\item{timeAd}{approximate length of adult life stage}
}
\description{
Calls \code{\link{set_initialPopulation_Patch}} to initialize a steady-state
population distribution. Populations are then rounded to integer values.
}
