% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGDrivE-Auxiliary.R
\name{splitOutput}
\alias{splitOutput}
\title{Split Output by Patch}
\usage{
splitOutput(directory, multiCore = FALSE)
}
\arguments{
\item{directory}{Directory where output was written to; must not end in path seperator}

\item{multiCore}{Write output using multiple cores? Default is FALSE}
}
\description{
Split output into multiple files by patches.
}
